kie-aries-blueprint: minor fixes. Correcting the incorrect object metadata mapping for kmodule.
